use sass helper method for asset url
we were trying to get the asset url manually using sprocket and actioncontroller helpers but the url generated by these helpers was not recognized by sprockets as a dependant link of css files and the svgs were not appended a digest. This commit allows voltron-svg use sass helper methods to get asset url and this allows sprocket append a digest to svg urls.